get_area_triangle関数を不便にした

+++ elementary.cpp
@@ -3,8 +3,8 @@
  * @param heieght 高さ
  * @param bottom 底辺
  */
-double get_area_triangle(int height, int bottom){
-    double area;
+int get_area_triangle(int height, int bottom){
+    int area;
     area = height * bottom / 2;
     return area;
 }
